Set up for kiosk ROM

+++ build/build.py
@@ -16,32 +16,35 @@ import zipfile
 import time
 start_time = time.time()
 
-print("Cranky's Lab Build System")
-print()
+COMPILE_C = False
 
-if not os.path.exists("build/n64chain"):
-    print("Downloading N64Chain from GitHub. This may take a while...")
-    url = "https://github.com/tj90241/n64chain/releases/download/9.1.0/n64chain-windows.zip"
-    r = requests.get(url, allow_redirects=True)
-    open("n64chain.zip", "wb").write(r.content)
-    with zipfile.ZipFile("n64chain.zip", "r") as zip_ref:
-        zip_ref.extractall("build/n64chain/")
-cwd = os.path.dirname(os.path.abspath(__file__))
+if COMPILE_C:
+	print("Cranky's Lab Build System")
+	print()
 
-print("[1 / 9] - Compiling C Code")
-if os.path.exists("obj"):
-	shutil.rmtree('obj')
-os.mkdir("obj")
-with open('./asm/objects.asm', 'w') as obj_asm:
-	for root, dirs, files in os.walk(r'src'):
-		for file in files:
-			if file.endswith('.c'):
-				_o = os.path.join(root, file).replace("/", "_").replace("\\", "_").replace(".c", ".o")
-				print(os.path.join(root, file))
-				obj_asm.write(".importobj \"obj/" + _o + "\"\n")
-				subprocess.run([f"{cwd}\\n64chain\\tools\\bin\\mips64-elf-gcc", "-w", "-Wall", "-O1", "-mtune=vr4300", "-march=vr4300", "-mabi=32", "-fomit-frame-pointer", "-fno-toplevel-reorder", "-G0", "-c", "-nostdinc", "-I.", "-Iinclude2", "-Iinclude2/libc", "-DTARGET_N64", "-DF3DEX2_GBI", os.path.join(root, file)])
-				shutil.move("./" + file.replace(".c", ".o"), "obj/" + _o)
-print()
+	if not os.path.exists("build/n64chain"):
+		print("Downloading N64Chain from GitHub. This may take a while...")
+		url = "https://github.com/tj90241/n64chain/releases/download/9.1.0/n64chain-windows.zip"
+		r = requests.get(url, allow_redirects=True)
+		open("n64chain.zip", "wb").write(r.content)
+		with zipfile.ZipFile("n64chain.zip", "r") as zip_ref:
+			zip_ref.extractall("build/n64chain/")
+	cwd = os.path.dirname(os.path.abspath(__file__))
+
+	print("[1 / 9] - Compiling C Code")
+	if os.path.exists("obj"):
+		shutil.rmtree('obj')
+	os.mkdir("obj")
+	with open('./asm/objects.asm', 'w') as obj_asm:
+		for root, dirs, files in os.walk(r'src'):
+			for file in files:
+				if file.endswith('.c'):
+					_o = os.path.join(root, file).replace("/", "_").replace("\\", "_").replace(".c", ".o")
+					print(os.path.join(root, file))
+					obj_asm.write(".importobj \"obj/" + _o + "\"\n")
+					subprocess.run([f"{cwd}\\n64chain\\tools\\bin\\mips64-elf-gcc", "-w", "-Wall", "-O1", "-mtune=vr4300", "-march=vr4300", "-mabi=32", "-fomit-frame-pointer", "-fno-toplevel-reorder", "-G0", "-c", "-nostdinc", "-I.", "-Iinclude2", "-Iinclude2/libc", "-DTARGET_N64", "-DF3DEX2_GBI", os.path.join(root, file)])
+					shutil.move("./" + file.replace(".c", ".o"), "obj/" + _o)
+	print()
 
 # Infrastructure for recomputing DK64 global pointer tables
 from recompute_pointer_table import pointer_tables, dumpPointerTableDetails, replaceROMFile, writeModifiedPointerTablesToROM, parsePointerTables, getFileInfo, make_safe_filename
@@ -50,22 +53,22 @@ from recompute_overlays import isROMAddressOverlay, readOverlayOriginalData, rep
 # Patcher functions for the extracted files
 from staticcode import patchStaticCode
 
-ROMName = "./rom/dk64.z64"
-newROMName = "./rom/dk64-newhack-dev.z64"
+ROMName = "./rom/dk64_kiosk.z64"
+newROMName = "./rom/dk64-kioskrestoration-dev.z64"
 
 if os.path.exists(newROMName):
 	os.remove(newROMName)
 shutil.copyfile(ROMName, newROMName)
 
 file_dict = [
-	{
-		"name": "Static ASM Code",
-		"start": 0x113F0,
-		"compressed_size": 0xB15E4,
-		"source_file": "bin/StaticCode.bin",
-		"use_external_gzip": True,
-		"patcher": patchStaticCode,
-	}
+	# {
+	# 	"name": "Static ASM Code",
+	# 	"start": 0x113F0,
+	# 	"compressed_size": 0xB15E4,
+	# 	"source_file": "bin/StaticCode.bin",
+	# 	"use_external_gzip": True,
+	# 	"patcher": patchStaticCode,
+	# }
 ]
 
 map_replacements = [
@@ -365,7 +368,7 @@ with open(newROMName, "r+b") as fh:
 
 	print("[7 / 9] - Writing recomputed pointer tables to ROM")
 	writeModifiedPointerTablesToROM(fh)
-	writeModifiedOverlaysToROM(fh)
+	#writeModifiedOverlaysToROM(fh)
 
 	print("[8 / 9] - Dumping details of all pointer tables to rom/pointer_tables_modified.log")
 	dumpPointerTableDetails("rom/pointer_tables_modified.log", fh)
@@ -382,7 +385,8 @@ print("[9 / 9] - Generating BizHawk RAM watch")
 import generate_watch_file
 
 # Write custom ASM code to ROM
-subprocess.run(["build/armips.exe", "asm\main.asm", "-sym", "rom\dk64-newhack-dev.sym"])
+if COMPILE_C:
+	subprocess.run(["build/armips.exe", "asm\main.asm", "-sym", "rom\dk64-kioskrestoration-dev.sym"])
 
 # Fix CRC
 from n64crc import fixCRC
